Successfully implemented left and right swipes.
Working player script PlayerMovement to add constraint on the rigid body component to prevent the player from falling or toppling over.
Enhanced the TouchManager script to allow to swipe movement by implementing 4 touch directions in the script RightMove(), LeftMove(), ForwardMove() and BackwardMove().
Enhanced the CameraController script to add motion smoothness to the camera.